+++ packages/web-app/src/modules/earn-views/components/EarningLineChart/components/index.tsx
@@ -1,0 +1,1 @@
+export * from './CustomizedXAxisTick'
