+++ packages/web-app/src/modules/earn-views/components/EarningLineChart/index.tsx
@@ -1,0 +1,1 @@
+export { EarningLineChart } from './EarningLineChart'
