+++ packages/web-app/src/modules/earn-views/components/index.tsx
@@ -1,5 +1,6 @@
+export * from './EarnSectionHeader'
 export * from './EarningFrequentlyAskedQuestions'
 export * from './EarningHistory'
+export * from './EarningLineChart'
 export * from './EarningSummary'
-export * from './EarnSectionHeader'
 export * from './LatestRewardsRedeemed'
